Refactored dlq creation
